requirements.txt
flask>=2.0.1
flask-sqlalchemy>=2.5.1
flask-login>=0.5.0
numpy>=1.24.0
pandas>=1.5.0
requests>=2.26.0
python-dotenv>=0.19.0
geopy>=2.2.0
meshtastic>=1.2.0
flask-socketio>=5.1.1
geopandas>=0.10.2
rtree>=1.0.0
shapely>=1.8.0
scikit-learn>=1.0.0
